Set up test for admin merchant index dashboard
